Metodos_de_Classificacao/Classification_Main/requirements.txt
Pillow
scikit-learn
opencv-python
numpy